refactor(major): file organization, build process
