+++ .eslintignore
@@ -1,3 +1,3 @@
-./chrome/**/*
-./firefox/**/*
-./node_modules/**/*
+chrome/*
+firefox/*
+node_modules/*
