--- firefox/data/inject.css
@@ -1,30 +1,0 @@
-.contrib-column {
-    text-align: center;
-    font-size: 11px;
-    padding: 15px 0px;
-    border-left: 1px solid rgb(221, 221, 221);
-    border-top: 1px solid rgb(221, 221, 221); 
-}
-.contrib-column-first {
-    border-left: 0px;
-}
-.table-column {
-    display: table-cell;
-    width: 1%;
-    padding-right: 10px;
-    padding-left: 10px;
-    vertical-align: top;
-}
-.text-muted {
-    color: rgb(118, 118, 118);
-}
-.contrib-number {
-    font-weight: 300;
-    line-height: 1.3em;
-    font-size: 24px;
-    display: block;
-    color: rgb(51, 51, 51);
-}
-.text-muted {
-    color: rgb(118, 118, 118);
-}
