+++ package.json
@@ -4,13 +4,10 @@
   "description": "Get back longest and current streak on Github prifiles",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint .; exit 0",
-    "install": "npm i -g babel-cli watch jpm",
-    "clean": "find build -regextype posix-egrep -regex \".*\\.(zip|xpi|nex)$\" -type f -delete && rmdir build",
-    "build:js": "babel ./src/inject.js --presets es2015 | tee ./firefox/data/inject.js > ./chrome/src/inject/inject.js ",
-    "build:css": "echo ./firefox/data/ ./chrome/src/inject/ | xargs -n 1 cp ./src/inject.css",
-    "build": "npm run build:js && npm run build:css",
-    "build:watch": "watch \"npm run build\" ./src/",
+    "lint": "eslint .",
+    "prebuild": "npm run lint && rimraf chrome/src/ firefox/data/ build",
+    "build": "webpack",
+    "build:watch": "webpack --watch",
     "package": " mkdir -p build && cd ./firefox && jpm xpi && mv *.xpi ../build && cd .. && zip -r ./build/chrome.zip ./chrome && cd ./build && cp chrome.zip chrome.nex"
   },
   "repository": {
@@ -29,8 +26,14 @@
   },
   "homepage": "https://github.com/Naramsim/GithubOriginalStreak#readme",
   "devDependencies": {
+    "babel-core": "^6.9.1",
+    "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.9.0",
+    "css-loader": "^0.23.1",
     "eslint": "^2.12.0",
-    "watch": "^0.18.0"
+    "jpm": "^1.0.7",
+    "rimraf": "^2.5.2",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.13.1"
   }
 }
